Complete module4-task1
Создал модули util.js, data.js и перенес в них код из main.js. Сделал все функции в util.js и data.js стрелочными, чтобы соответствовать критерию Д5 (все функции объявлены единообразно)
